requirements.txt
# opencv-python  # уже содержится в opencv-contrib-python, поэтому не обязателен
opencv-contrib-python==4.6.0.66  # именно эта версия. 4.7.0.72 например давала ошибку.
flask
celery
pillow